Добавлена поддержка вещественных чисел

+++ src/Main.java
@@ -7,22 +7,21 @@ public class Main {
         System.out.print("Введите число: ");
         Scanner input = new Scanner (System.in);
         //присвоение
-        int chislo1 = input.nextInt();
+        double chislo1 = input.nextInt();
 
         //      Второе число
         //запрос
         System.out.print("Введите второе число: ");
         Scanner input2 = new Scanner (System.in);
         //присвоение
-        int chislo2 = input2.nextInt();
+        double chislo2 = input2.nextInt();
 
         //      Действие
         //запрос
         System.out.print("Введите номер действия: ");
         Scanner input3 = new Scanner (System.in);
         int znak = input3.nextInt();
-
-        //сложние
+        //сложение
         if(znak==1){
             System.out.print("Сумма чисел: " + (chislo1 + chislo2));
         }
@@ -38,6 +37,5 @@ public class Main {
         if(znak==4){
             System.out.print("Разность чисел: " + (chislo1 - chislo2));
         }
-
     }
 }
